# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IT-Ticket API
  version: 1.1.0
  description: >
    REST-API för IT-Ticket — ett ärendehanteringssystem för intern IT-support på Prefabmästarna.


    **Leverans:** rullande från `main` (en instans per deployment).
    Versionen ovan spårar API-kontraktet, inte appens git-tagg.


    ## API coverage

    This spec documents the full API surface: 160 operations
    across every route file — health/CSRF, auth, tickets, categories,
    contacts, companies, kb, public, users, API keys,
    settings, comments, attachments, shares, links, checklists, checklist
    templates, template fields, tags,
    reports, backup, push notifications, webhooks, and inbound email.


    **Prose reference:** `docs/API.md` — generated from the actual Express
    routers in `server/src/routes/` and kept up to date on route changes. Use
    it as a narrative cross-check alongside this machine-readable spec.


    ## Autentisering

    Två sätt att autentisera, båda via `Authorization`-headern:

    - **Session (JWT):** `Authorization: Bearer <jwt>`. Access-token gäller 15 min;
      en HttpOnly refresh-cookie (path `/api/auth`) förnyar den via
      `POST /api/auth/refresh`. **Muterande anrop (POST/PUT/PATCH/DELETE) kräver
      även CSRF-token** i `x-csrf-token` (hämtas från `GET /api/csrf-token`).
    - **API-nyckel:** `Authorization: Bearer itk_live_...`. Nycklar har scope
      `read`, `read,write` eller — för administrativa endpoints — även `admin`.
...


...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: Lista kategorier
      responses:
        "200":
          description: Kategorier
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      tags: [Categories]
      summary: Skapa kategori (admin)
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
      responses:
        "201":
          description: Skapad kategori
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    put:
      tags: [Categories]
      summary: Uppdatera kategori (admin)
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
      responses:
        "200":
          description: Uppdaterad kategori
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Categories]
      summary: Radera kategori (admin)
      security:
        - bearerAuth: []
          csrfToken: []
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /kb/portal/{token}/categories:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string } }
    get:
      tags: [Knowledge base]
      summary: Lista portalens kategorier
      description: Oautentiserad, olistad och rate-limited. Endast kategorier med publicerade artiklar returneras.
      security: []
      responses:
        "200":
          description: Kategorier med artikelantal
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/KbPortalCategory" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/Error" }
...
  /kb/categories:
    get:
      operationId: kbListCategories
      tags: [Knowledge base]
      summary: List KB categories
      description: Any authenticated user. `article_count` counts published articles only.
      responses:
        "200":
          description: Categories ordered by position then name
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/KbCategory" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: kbCreateCategory
      tags: [Knowledge base]
      summary: Create a KB category (admin)
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KbCategoryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KbCategory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /kb/categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    put:
      operationId: kbUpdateCategory
      tags: [Knowledge base]
      summary: Update a KB category (admin)
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KbCategoryInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KbCategory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: kbDeleteCategory
      tags: [Knowledge base]
      summary: Delete a KB category (admin)
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /categories/reorder:
    put:
      operationId: categoriesReorder
      tags: [Categories]
      summary: Reorder categories (admin)
      description: The position of each id becomes its index in `ids`.
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  minItems: 1
                  items: { type: string }
      responses:
        "200":
          description: Categories in the new order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ───────────────────────── Auth: audit log ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Session-JWT: `Authorization: Bearer <jwt>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API-nyckel: `Authorization: Bearer itk_live_...` (scope read/write/admin; admin-skyddade endpoints kräver `admin` även när ägaren är administratör). Undantagen från CSRF."
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Krävs på muterande sessions-anrop. Hämtas från `GET /api/csrf-token`.
  responses:
    Error:
      description: Fel
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Message:
      description: OK med meddelande
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
    Unauthorized:
      description: Ej autentiserad
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Saknar behörighet (t.ex. admin krävs eller fel scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Hittades inte
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        label: { type: string }
        position: { type: integer }
        created_at: { type: string, format: date-time }
...
    KbPortalCategory:
      type: object
      required: [id, name, article_count]
      properties:
        id: { type: string }
        name: { type: string }
        color: { type: string, nullable: true }
        article_count: { type: integer, minimum: 1 }
...
    KbCategory:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        color: { type: string, nullable: true }
        position: { type: integer }
        created_at: { type: string, format: date-time }
        article_count: { type: integer, description: Published articles only. Present on the list endpoint. }
    KbCategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        color: { type: string, nullable: true }
...